moved GCSRegistryStore to gcp
